backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas==2.2.2
qdrant-client==1.11.0
groq>=1.1.0
cerebras_cloud_sdk>=1.0.0
openai>=1.30.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
pydantic==2.8.0
